[ADD] A third state in onConnectionChanged, now we have : isConnected, isConnectedorConnecting, on disconnection
NOTES :
a) I think _bWasConnected should not be in onConnectionChanged because, the connection might fail and the boolean might indicate incorrect stage (to explore)
b) I tried to do launchServicesRequestPeers only if the tablettes were not connecting or connected, because,  why should they launch a discovery if they are connected or in the procces of connection ? But, this leads to a crash.

TESTS (31/01/2018) using the Nantes game:

In both cases, we did not have a "disconnection message" while launching the vidéo.

1. Cenuschi tablets :
       After separation, they connected if the other tablet accepted the invitation, if cancel the invitation and if not response is given.
      The 3 windows appeared (disconnection, temptative of connection, connection) in one tablet, in the other one, I had to push "next". Seems that in the other tablet (the one that sends the invitation), it needs to push "next" to popup the window

2. Nantes smartphones :
     1. It connects, after several trials

TESTING (31/01/2018) using Cernuschi game

1. Cenuschi tablets pass ok the sending photo.

ISSUES :
    When connecting between devices, the Cernuschi the  orange message (of patientez... ) is still present, even if the tablets are connected. The logcat show "listen socket server" as it was connected..