[Leak Fixed] Avoid connecting task when already connected
[Test - 05/02/2018]
- We ran ludomuse on ludo and cernoushi tablets simultaneously.
- We went away with two tablets from differents couple and then came back to test reconnexion.
- On Ludo tablet, reconnexion appear without notification message
- On Cernoushi, reconnexion appear after a notification message. We tested the reconnexion were effective,
after cancelling, authorizing, or waiting for notification message to disappear on the screen (up to 30 s).
The three previous conditions were tested at least 3 times.

[Issue]
The following procedure may be used when connexion failed for unknown reason :
- try to on/off the screen display (without rebooting tablet)
- try to on/off WIFI on both tablet
(could be due automatically or left to the user through an UI button)